fix: all 18 audit findings from Java terminal comparison
Closes #3

CRITICAL (3):
- #1: Contract ID now FIT-decoded as first field (was raw BE i32)
- #2: Delta off-by-one fixed (exclude contract_id from tick delta)
- #3: Added null_value (field 5) to DataValue oneof in proto

HIGH (4):
- #4: Trade field mapping correct after #1 fix (verified)
- #5: Added "client": "terminal" to QueryInfo.query_parameters
- #6: Delta state cleared on START/STOP (market open/close)
- #7: Dynamic concurrency from subscription tier (2^max_ordinal)

MEDIUM (5):
- #8: Unknown compression algo returns error (was silent NONE)
- #9: Empty stream returns empty DataTable (was Error::NoData)
- #10: AuthUser now has 4 per-asset subscription fields
- #11: OHLCVC-from-trade documented as TODO (needs live verification)
- #12: Column lookup warns on missing headers (was silent fallback)

LOW (6):
- #13: Added gRPC initialFlowControlWindow (65536 bytes)
- #14: ROW_SEP unconditionally resets idx to SPACING
- #15: readChanges return documented (0 vs -1, equivalent)
- #16: Credential length sign-extension matches Java putShort((byte))
- #17: Flush only on PING frames (was every frame)
- #18: Ping loop starts after 2000ms delay (was immediate)

+++ crates/thetadatadx/src/fpss/protocol.rs
@@ -363,7 +363,13 @@ impl std::error::Error for ContractParseError {}
 pub fn build_credentials_payload(username: &str, password: &str) -> Vec<u8> {
     let user_bytes = username.as_bytes();
     let pass_bytes = password.as_bytes();
-    let user_len = user_bytes.len() as u16;
+
+    // Match Java's `putShort((byte)len)` behavior: the length is first narrowed
+    // to a byte (i8), then sign-extended to a short (i16). For lengths 0-127
+    // this is identical to a plain u16 cast. For lengths 128-255 the sign
+    // extension sets the high byte to 0xFF. In practice usernames are always
+    // <128 bytes, but we match the exact wire encoding for correctness.
+    let user_len = user_bytes.len() as i8 as i16;
 
     // 1 (version) + 2 (user_len) + user + pass
     let mut buf = Vec::with_capacity(3 + user_bytes.len() + pass_bytes.len());
